+++ NodeDependencyLookup/Editor/Caches/AssetDependencyCache/AssetTraverser/AssetSerializedPropertyTraverser.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Reflection;
 using UnityEditor;
 using UnityEngine;
@@ -12,6 +13,9 @@ namespace Com.Innogames.Core.Frontend.NodeDependencyLookup
 	/// </summary>
 	public class AssetSerializedPropertyTraverser : AssetTraverser
 	{
+		private static int TraversablePropertyTypes = (1 << (int)SerializedPropertyType.ObjectReference) |
+			(1 << (int)SerializedPropertyType.Generic) | (1 << (int)SerializedPropertyType.ManagedReference);
+
 		private readonly Stack<PathSegment> pathSegmentStack = new Stack<PathSegment>();
 
 		private readonly HashSet<Type> excludedTypes = new HashSet<Type>
@@ -65,24 +69,37 @@ namespace Com.Innogames.Core.Frontend.NodeDependencyLookup
 			unsafeModeMethod.SetValue(property, true);
 			property.Next(true);
 
-			SerializedPropertyType propertyType;
+			var traverseChildren = false;
 
 			do
 			{
-				propertyType = property.propertyType;
+				var propertyType = property.propertyType;
 
-				if (propertyType != SerializedPropertyType.ObjectReference &&
-				    propertyType != SerializedPropertyType.Generic &&
-				    propertyType != SerializedPropertyType.ManagedReference)
+				if ((TraversablePropertyTypes & (1 << (int)propertyType)) == 0)
 				{
 					continue;
 				}
+
+#if UNITY_6000_3_OR_NEWER
+				if (property.depth > 256)
+				{
+					var assetPath = AssetDatabase.GetAssetPath(searchContext.Asset);
+					var fileName = Path.GetFileName(assetPath);
+					Debug.LogWarning(
+						$"[ARV] Reached max depth in {fileName}. " +
+						$"Skipping [{property.propertyPath.Substring(0, 64)}]", searchContext.Asset);
+					break;
+				}
+#endif
 
 				if (!onlyOverriden || property.prefabOverride)
 				{
 					TraverseProperty(searchContext, property, propertyType, property.propertyPath, stack);
 				}
-			} while (property.Next(propertyType == SerializedPropertyType.Generic || propertyType == SerializedPropertyType.ManagedReference));
+
+				traverseChildren = propertyType == SerializedPropertyType.Generic ||
+					propertyType == SerializedPropertyType.ManagedReference;
+			} while (property.Next(traverseChildren));
 
 			serializedObject.Dispose();
 		}